requirements.txt
pywin32-308 xlwings-0.33.6 (pip install xlwings)

PyPDF2-3.0.1 (pip install PyPDF2)

# first run the main script then run the conversion script